### 2019-10-16 #1815 @whedon
[REVIEW]: MDL Suite: A language, generator and compiler for describing mazes
**Submitting author:** @akashnag (<a href="http://orcid.org/0000-0003-0363-2698">Akash Nag</a>)
**Repository:** <a href="https://github.com/akashnag/MDL-Suite" target ="_blank">https://github.com/akashnag/MDL-Suite</a>
**Version:** 1.0
