### 2019-10-16 Akash-Nag/MDL-Suite#2 @rreinecke
Mention contributors in README
You did acknowledge someone in your paper please consider adding a list of contributors in your README as well.
