### 2019-10-16 Akash-Nag/MDL-Suite#3 @rreinecke
Compile guide and version statement
Please add a description with which versions of java your build is compatible with and consider adding a small build instruction.

With the default Ubuntu JRE I'm getting:

`Exception in thread "main" java.lang.UnsupportedClassVersionError: mdlg/MDLG has been compiled by a more recent version of the Java Runtime (class file version 54.0), this version of the Java Runtime only recognizes class file versions up to 52.0
`
