+++ README.md
@@ -10,6 +10,22 @@ This suite contains two tools:
 
 Both the generator and compiler are written in Java. The pre-built JARs are available in the `build` directory and executing them requires a JVM (Oracle Java 10, version 10.0.2) to be installed. There are no other dependencies for this software. For building from source, any JVM (version 8 onwards) can be used and the Java source files are available in the `mdlc` and `mdlg` directories.
 
+To compile from source (without a JAR) from the parent directory:
+
+```shell
+javac mdlc/MDLC.java
+javac mdlg/MDLG.java
+```
+
+To execute the tools:
+
+```shell
+java mdlc.MDLC <parameters>
+java mdlg.MDLG <parameters>
+```
+
+where the `<parameters>` refer to the list of input files needed to be passed as arguments.
+
 ## Using the MDL generator to generate mazes
 
 The MDL generator tool is named **MDLG** and can be invoked from the command-line. It takes 2 or more parameters. The first parameter is a **configuration file** which is typically named ``maze.config``. This will contain the various settings for the mazes that will be generated by MDLG. The names of these settings can be found in the section on the MDL language below.
@@ -210,6 +226,14 @@ $ java -jar mdlc.jar maze2.maze
 If the command executes successfully, we will be able to see that a file named ``maze2.png`` has been generated as follows (the exact output will vary as the maze is generated dynamically and randomly):
 
 ![Output from Example-2](./examples/example2.png)
+
+## Contact
+
+For any queries regarding this software or how to use it, you can contact the author directly:
+
+**Akash Nag**
+
+E-mail: `nag[dot]akash[dot]cs[at]gmail.com`.
 
 ## Acknowledgements
 
