### 2019-10-16 Akash-Nag/MDL-Suite#3 @rreinecke
Thanks! Please add how to build the jar as well (just as convenience) because `java mdlg.MDLG` is producing: `java mdlg.MDLG
Maze Description Language Generator - v1.0
by Akash Nag. This software is open-source and is distributed under Artistic License 2.0.
Usage:
	 java -jar mdlg.jar <config-file> <output-files-in-mdl-format>
`
