### 2019-10-16 Akash-Nag/MDL-Suite#3 @rreinecke
Furthermore: `java mdlg.MDLG example/maze.config maze1 maze2 maze3` is yielding `java.io.FileNotFoundException: example/maze.config (No such file or directory)`
