### 2019-10-16 Akash-Nag/MDL-Suite#5 @rreinecke
Add tests
As also mentioned in  #4 consider adding test cases. This allows me to further evaluate the functionality of the code and other to contribute to the project and overall maintain the quality of the software.
