### 2019-10-16 Akash-Nag/MDL-Suite#3 @akashnag
Sorry, it should be examples/maze.config instead of example/maze.config
