+++ mdlg/MDLG.java
@@ -1,17 +1,54 @@
+/**
+ * Maze Description Language Generator
+ * Distributed under Artistic License 2.0
+ * 
+ * Generates random mazes in MDL format using the recursive division method
+ * 
+ * @author	Akash Nag
+ * @version	1.0
+ */
+
 package mdlg;
 
 import java.io.*;
 
+/**
+ * This is the main class handling the user-parameters, reading configuration files
+ * and calling appropriate maze generation functions.
+ */
 public class MDLG
 {
+	/**
+	 * The current version of the MDL generator
+	 */
 	public static final String VERSION = "1.0";
+
+	/**
+	 * The usage string that will be displayed if the program is invoked
+	 * without parameters or with incorrect parameters
+	 */
 	private static final String INFO = "Maze Description Language Generator - v" + VERSION +
 										"\nby Akash Nag. This software is open-source and is distributed under Artistic License 2.0." +
 										"\nUsage:\n\t java -jar mdlg.jar <config-file> <output-files-in-mdl-format>";
 										
+	/**
+	 * The width of the maze (in number of cells) that will be generated
+	 */
 	private static int width;
+
+	/**
+	 * The height of the maze (in number of cells) that will be generated
+	 */
 	private static int height;
 
+	/**
+     * This is the main method, and it processes the command-line arguments
+     * 
+     * @param args 			The name of the configuration file followed by one or more output file names
+     * @return 				Nothing
+     * @exception Exception On input error or error in the syntax of the configuration file.
+     * @see 				Exception
+     */
 	public static void main(String args[]) throws Exception
 	{
 		if(args.length < 2)
@@ -34,12 +71,28 @@ public class MDLG
 		}
 	}
 
+	/**
+	 * Displays the version information and the list of parameters 
+	 * to be passed to the program at command-line
+	 * 
+	 * @param 	Nothing
+	 * @return 	Nothing
+	 */
 	private static void printUsage()
 	{
 		System.out.println(INFO);
 	}
 
-	private static String readConfiguration(String fileName) throws Exception
+	/**
+	 * Reads the configuration file passed as the first parameter to the program.
+	 * The configuration file contains settings that will be common to all generated mazes.
+	 * 
+	 * @param fileName			Name of the configuration file
+	 * @return 					A string containing a new-line-separated list of configurations
+	 * @exception IOException	On input error
+	 * @see 					IOException
+	 */
+	private static String readConfiguration(String fileName) throws IOException
 	{
 		StringBuffer sb = new StringBuffer();
 		BufferedReader br = new BufferedReader(new FileReader(fileName));
@@ -59,6 +112,15 @@ public class MDLG
 		return sb.toString();
 	}
 
+	/**
+	 * Merges and writes the configuration and generated maze into an output file in MDL format
+	 * 
+	 * @param maze				A boolean matrix representing the generated maze, indicating true for path or false for blocked cells
+	 * @param config			The list of configurations about the maze
+	 * @param outputFile		The name of the output file
+	 * @return 					Nothing
+	 * @exception IOException	On output error
+	 */
 	private static void writeMazeToFile(boolean maze[][], String config, String outputFile) throws IOException
 	{
 		BufferedWriter bw = new BufferedWriter(new FileWriter(outputFile));
@@ -67,6 +129,12 @@ public class MDLG
 		bw.close();
 	}
 
+	/**
+	 * Prepares a generated maze for writing to a file by converting it from boolean to a string in MDL format
+	 * 
+	 * @param maze		The generated maze
+	 * @return 			A string representation of the maze in MDL format
+	 */
 	private static String mazeToString(boolean maze[][])
 	{
[…]
