### 2019-10-17 Akash-Nag/MDL-Suite#3 @akashnag
The instructions for building the JARs are now available in the README. I have also added the documentation.
