### 2019-10-17 Akash-Nag/MDL-Suite#6 @rreinecke
Running examples fails
1.  Please add the correct path to the maze.config in the README
2.  `java -jar mdlg.jar examples/maze.config maze1 maze2 maze3
1/3: maze generated successfully
2/3: maze generated successfully
3/3: maze generated successfully
` works, but `java -jar mdlc.jar maze1
An error occurred during maze generation:
java.lang.StringIndexOutOfBoundsException: String index out of range: -1
java.lang.StringIndexOutOfBoundsException: String index out of range: -1
	at java.lang.String.substring(String.java:1967)
	at mdlc.MDLC.main(MDLC.java:22)
` also the files generated in the first step are not called maze1.maze as in the README example.
