### 2019-10-17 #1815 @rreinecke
@Akash-Nag please address the following issues https://github.com/Akash-Nag/MDL-Suite/issues/5 https://github.com/Akash-Nag/MDL-Suite/issues/6 . Especially the automated tests will help to understand your code better. Without them I'm currently not able to proceed with the review. Please also consider improving your code documentation further.
If you have questions please feel free to ask in the according issues.
