### 2019-10-18 c5a322f
Added tests
