+++ README.md
@@ -51,10 +51,10 @@ Both these JARs can now be executed as explained later in this readme.
 
 The MDL generator tool is named **MDLG** and can be invoked from the command-line. It takes 2 or more parameters. The first parameter is a **configuration file** which is typically named ``maze.config``. This will contain the various settings for the mazes that will be generated by MDLG. The names of these settings can be found in the section on the MDL language below.
 
-To generate mazes with MDLG, we therefore need to pass the path to the configuration file followed by the paths to 1 or more output files (without extensions). Note: if the output files already exist, they will be overwritten. The number of mazes generated will be the number of output file names specified. For example, to generate 3 mazes, named: maze1, maze2, and maze3, we could execute the following command (assuming that the ``maze.config`` file exists already in the same directory):
+To generate mazes with MDLG, we therefore need to pass the path to the configuration file followed by the paths to 1 or more output files (without extensions). Note: if the output files already exist, they will be overwritten. The number of mazes generated will be the number of output file names specified. For example, to generate 3 mazes, named: maze1, maze2, and maze3, we could execute the following command (assuming that the ``maze.config`` file exists in the examples directory):
 
 ```shell
-$ java -jar mdlg.jar maze.config maze1 maze2 maze3
+$ java -jar build/mdlg.jar examples/maze.config maze1.maze maze2.maze maze3.maze
 ```
 
 On executing the above command, 3 files: ``maze1.maze``, ``maze2.maze`` and ``maze3.maze`` will be generated, each containing a random maze as per the specifications in ``maze.config``.
@@ -66,10 +66,21 @@ The MDL Compiler is named MDLC and can be used to generate an image, or code in 
 The compiler takes 1 or more MDL (\*.maze) files as input, and generates an output corresponding to each file. In the above example, we had generated 3 maze description files. We can now use the compiler to generate 3 PNG images from these files, as follows:
 
 ```shell
-$ java -jar mdlc.jar maze1.maze maze2.maze maze3.maze
+$ java -jar build/mdlc.jar maze1.maze maze2.maze maze3.maze
 ```
 
 The above command will generate 3 maze output files, which can be either images or code depending on the settings specified inside the source files. To know how to set the output modes inside the source files, please refer to the section on the MDL language below.
+
+## Using the tests
+
+Automated maze generation scripts are available as Linux shell scripts in the `tests` directory. To execute the test `test-gen.sh`:
+
+```shell
+cd tests
+sudo sh test-gen.sh
+```
+
+The script will prompt the user to enter how many mazes to generate. If the user enters an integer like 5, five mazes (both .maze and .png files) will be generated inside the `tests/test_outputs` directory.
 
 ## MDL Language Documentation
 
