### 2019-10-18 Akash-Nag/MDL-Suite#6 @akashnag
The following changes have been made:
1. The exact command to execute the tools from the root directory have been included in the README
2. I have added an automated test utility available as a shell-script inside the tests directory
