### 2019-10-18 Akash-Nag/MDL-Suite#5 @akashnag
An automated test utility has been added as a shell script inside the tests directory
