### 2019-10-18 Akash-Nag/MDL-Suite#6 @akashnag
The tests generate the maze according to the configuration specified in the maze.config file. You can change these settings to obtain different results. e.g. The generated mazes are black-and-white by default. You can generate blue mazes (say) by adding the following setting at the end of maze.config:

[wall-color:rgb(0:0:255)]

After this change, re-executing the test will cause the wall colors to be blue.
