### 2019-10-21 Akash-Nag/MDL-Suite#5 @rreinecke
Please make yourself familiar with JUnit and add according automated tests to your code that allow us and future developers to verify the functionality of your code. The shell script is not sufficient.
