### 2019-11-02 #1815 @danielskatz
👋 @Akash-Nag - can you update us on this review and your progress in addressing the issues discussed above?
