### 2019-11-03 #1815 @Akash-Nag
All the issues raised by the reviewers have been resolved, except for the
Tests. Incorporating JUnit tests is difficult for a project of this type
since this is not a library that can be invoked. Instead, it is a tool
whose output is an image file. Obviously, checking whether the generated
maze-image conforms to the input specifications requires visual inspection
or outlandish AI-driven image recognition algorithms beyond the scope of
this project. For internal functions which do not generate the image
directly, it may be possible to write tests but the difficulty there lies
with the fact that most of these functions work with matrices, and again,
checking whether these maze matrices conform to the specifications will
require their own algorithms that are unknown to me. The code for the tests
will, in all likelihood, be larger and more complicated than the source
code itself. Therefore, I would request you to kindly proceed with the
review for this project as it stands now. For your reference, I have made
the following changes:
1. The readme now contains detailed installation and build instructions,
names of contributors, etc.
2. The JavaDocs has been made available
3. A shell script has been provided for automated generation of random
mazes, for testing purposes
