### 2019-11-04 #1815 @Akash-Nag
The specific version of Java with which to run this tool has already been
highlighted in the readme under the "Dependencies and Build Instructions"
section, namely: Oracle Java 10.0.2

On Mon, 4 Nov 2019, 05:20 André Leon S. Gradvohl, <notifications@github.com>
wrote:

> Regarding the docs, I think the author solved this issue. However, when I
> tried to run, I still get the following error:
>
> Exception in thread "main" java.lang.UnsupportedClassVersionError: mdlg/MDLG has been compiled by a more recent version of the Java Runtime (class file version 54.0), this version of the Java Runtime only recognizes class file versions up to 52.0
>         at java.lang.ClassLoader.defineClass1(Native Method)
>         at java.lang.ClassLoader.defineClass(ClassLoader.java:756)
>         at java.security.SecureClassLoader.defineClass(SecureClassLoader.java:142)
>         at java.net.URLClassLoader.defineClass(URLClassLoader.java:468)
>         at java.net.URLClassLoader.access$100(URLClassLoader.java:74)
>         at java.net.URLClassLoader$1.run(URLClassLoader.java:369)
>         at java.net.URLClassLoader$1.run(URLClassLoader.java:363)
>         at java.security.AccessController.doPrivileged(Native Method)
>         at java.net.URLClassLoader.findClass(URLClassLoader.java:362)
>         at java.lang.ClassLoader.loadClass(ClassLoader.java:418)
>         at sun.misc.Launcher$AppClassLoader.loadClass(Launcher.java:352)
>         at java.lang.ClassLoader.loadClass(ClassLoader.java:351)
>         at sun.launcher.LauncherHelper.checkAndLoadMain(LauncherHelper.java:495)
>
> Should I run it with a specific java version? If so, this should be
> highlighted in the README.md. as well as in the docs.
>
