### 2019-11-04 #1815 @rreinecke
> All the issues raised by the reviewers have been resolved, except for the Tests. Incorporating JUnit tests is difficult for a project of this type since this is not a library that can be invoked. Instead, it is a tool whose output is an image file. Obviously, checking whether the generated maze-image conforms to the input specifications requires visual inspection or outlandish AI-driven image recognition algorithms beyond the scope of this project. For internal functions which do not generate the image directly, it may be possible to write tests but the difficulty there lies with the fact that most of these functions work with matrices, and again, checking whether these maze matrices conform to the specifications will require their own algorithms that are unknown to me. The code for the tests will, in all likelihood, be larger and more complicated than the source code itself. Therefore, I would request you to kindly proceed with the review for this project as it stands now. For your reference, I have made the following changes: 1. The readme now contains detailed installation and build instructions, names of contributors, etc. 2. The JavaDocs has been made available 3. A shell script has been provided for automated generation of random mazes, for testing purposes
> […](#)
> On Sat, 2 Nov 2019, 19:25 Daniel S. Katz, ***@***.***> wrote: wave @Akash-Nag <https://github.com/Akash-Nag> - can you update us on this review and your progress in addressing the issues discussed above? — You are receiving this because you were mentioned. Reply to this email directly, view it on GitHub <#1815?email_source=notifications&email_token=ALUWTRYXNIHROLNTRD524ILQRWBGRA5CNFSM4JBKNNM2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEC44OOQ#issuecomment-549046074>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/ALUWTR5C3PRR57OB2Z23IZDQRWBGRANCNFSM4JBKNNMQ> .

@drvinceknight @danielskatz I cannot confirm the statement given by @Akash-Nag . Yes all other concerns except the community guidelines and tests have been addressed. However, I do not agree with the assessment that unit tests are impossible and I find it bewildering that our suggestions as reviewers are fully ignored. I understand that JOSS has a very constructive review process, thus I opened an issue more than two weeks ago suggesting unit tests and that I would be happy to provide additional inputs if necessary. This was completely ignored by @Akash-Nag  till the statement above was given. As a computer scientist I cannot agree that test are not possible. Leaving the image generation aside, the software contains a maze generator that is able to produce random mazes. It is unclear to me if they are always solvable? If so, how do you prove that? Can they also be unsolvable? It seems you are using recursive-division to generate them. Then at least test the method that is implementing that algorithm. @Akash-Nag writes that the software is intended to be used for research. IF that is the case it needs to be clear how the software works. You can only ensure this by providing test cases that verify this behavior.
