### 2019-11-05 #1815 @rreinecke
> I apologize for any misunderstanding that may have occurred. I never intended to say that tests would be impossible, but only very difficult given the type of the project. Furthermore, I was not sure which aspects of the project could be actually tested as I was sure that neither the maze image generation nor the path markers could be feasibly tested to ensure they conform to the specifications. Due to this confusion I was not able to reply to the issue, but in restrospect I think I should have asked for more clarification, and I am very sorry for that. However, now that you mention about the proof of solvability of the generated mazes, I think there will be two ways of going about it. Firsly, there can be a formal proof of the correctness of the algorithm in the paper itself, but I think the format of JOSS papers will not be able to accomodate that. Secondly, I can include a function which, given a generated maze matrix, will print out any one path (as a list of coordinates) from the entrance to the exit, or alternately, a test which will return true/false depending on whether or not a path exists. Kindly state which do you want.
> […](#)
> On Mon, 4 Nov 2019, 23:12 Vince Knight, ***@***.***> wrote: @Akash-Nag <https://github.com/Akash-Nag> I fully agree with @rreinecke <https://github.com/rreinecke> and do not feel that tests are impossible. Is this something you will be able to address? — You are receiving this because you were mentioned. Reply to this email directly, view it on GitHub <#1815?email_source=notifications&email_token=ALUWTR4NUO5LSGGWS5SVVN3QSBNHJA5CNFSM4JBKNNM2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEDADEYI#issuecomment-549466721>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/ALUWTRZTHEIKXYDJ2ZPJURDQSBNHJANCNFSM4JBKNNMQ> .

@Akash-Nag  please see https://github.com/Akash-Nag/MDL-Suite/issues/5 for reply and further discussion.
