### 2019-11-05 Akash-Nag/MDL-Suite#5 @rreinecke
> Firsly, there can be a formal proof of the correctness of the algorithm in the paper itself, but I think the format of JOSS papers will not be able to accomodate that. Secondly, I can include a function which, given a generated maze matrix, will print out any one path (as a list of coordinates) from the entrance to the exit, or alternately, a test which will return true/false depending on whether or not a path exists.

You are right a proof is not necessary and I suppose there is already a paper on that? did you cite it? Maybe you could add the reference in the docstring as well

Concerning the the test: Reading your suggestion I'm guessing that you never wrote uni tests before. Which is ok but you should familiarize yourself with the concept (see also comment above).
Looking at  https://github.com/Akash-Nag/MDL-Suite/blob/master/mdlg/RandomMazeGenerator.java#L63 there is a lot of potential there for tests and also much more error handling. Possible test cases that come to my mind (you should think of more yourself and how to verify them):

*  What happens if all parameters are zero?
*  Are there edge cases were the method is not working as intended? e.g. see above think about raising an exception instead of just returning. You should test for correct error handling afterwards
* what about a very thin maze? Is it possible to basically have 1 million columns and 2 rows? Is it still generating something useful? You should test this.
*  Additionally to edge cases it would be helpful to just randomly generate a thousand mazes and test if there is a possible path through

All implemented in JUnit tests. That would greatly help to verify that your code is actually doing what you promise and help future development!
