### 2019-11-28 #1815 @drvinceknight
Thanks, I'll wait to hear from them to confirm before I close this.

@openjournals/jose-eics can I confirm the process for closing ("rejecting for now") a submission? Is it as simple as closing this issue or is there some @whedon magic I should use instead?
