### 2019-11-28 #1815 @labarba
You can add the "paused" label
