### 2019-11-28 #1815 @rreinecke
@drvinceknight  for tracking completed reviews via e.g. publons this labeling is not really satisfying. If this project is never revived this does not count as a review. I'm very happy to review for JOSS anytime but it would be nice if you guys would appreciate that work with a completed review mail that at least allows to track that effort.
