### 2019-11-29 #1815 @Akash-Nag
@drvinceknight kindly advise regarding the process of resubmission. I'm sorry for not being able to make the necessary changes right now as I am very busy with too many unforeseen assignments. I apologize for the inconvenience caused. Please advise whether I should withdraw the paper (from JOSS website) and then resubmit later? Or, can/should this review be just paused, etc. in some way?
