### 2019-11-30 #1815 @drvinceknight
@rreinecke I understand completely. Thank you again for your time and work reviewing the paper. It's greatly appreciated.

@rreinecke and @Akash-Nag I am just clarifying the process myself via email, I will get back to you here as soon as I'm confident in the correct path forward.
