### 2019-12-02 #1815 @drvinceknight
Hi @rreinecke and @Akash-Nag.

Thank you for your patience.

I completely realise the effort and need for recognition of said effort of reviewers @rreinecke so once again; thank you and hopefully I can explain the options available so that we can choose the best one that importantly works for @Akash-Nag and the paper.

Option 1: The paper is withdrawn. This is the best option if @Akash-Nag is unlikely to come back to this in the next 3 months or so.

Option 2: We leave the paused option - this is best if @Akash-Nag is likely to come back to this in the next 3 months or so. @rreinecke I can set @whedon up to set a reminder so that we all do not forget, in which case in 3 months time we can revisit and if @Akash-Nag is unlikely to make progress we can go with option 1 at that stage.

Which way forward is preferable at this stage?
