### 2019-12-03 #1815 @drvinceknight
Assuming this is fine with you @rreinecke then the review is marked as pause. I will set a reminder for all of us of 3 months.
