### 2020-02-04 05c8ee1
Added JUnit Tests
