### 2020-02-04 #1815 @Akash-Nag
Hi,

My paper/project was on pause as you might remember because I was not
able to complete the requirements then. I have now updated my
repository with JUnit tests, but I am not sure how to change the
status of the project. Therefore, I am replying to this email.
@rreinecke Could you please check if these were the tests you wanted,
and if anything more is required?
