### 2020-02-05 #1815 @drvinceknight
I have removed the pause label.
