### 2020-02-05 Akash-Nag/MDL-Suite#7 @rreinecke
JOSS review: add community guidelines
Please add contribution guidelines (see other github projects for examples) that allows me to finish this part of the review:

"Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support "
