### 2020-02-05 Akash-Nag/MDL-Suite#8 @rreinecke
JOSS review: state of the field
Please add the following to the paper:

"State of the field: Do the authors describe how this software compares to other commonly-used packages?"
