+++ paper/paper.bib
@@ -13,3 +13,35 @@
   url = {http://www.micromouseonline.com/micromouse-book/history/},
   urldate = {2019-10-11}
 }
+
+@online{OppenheimJ:2018,
+  title = {GitHub - oppenheimj / maze-generator},
+  year = 2018,
+  url = {https://github.com/oppenheimj/maze-generator/},
+  urldate = {2020-02-06}
+}
+
+@online{MazeGenerator:2019,
+  title = {Maze Generator},
+  year  = 2019,
+  url   = {http://www.mazegenerator.net/},
+  urldate = {2020-02-06}
+}
+
+@online{Wolfram:2011,
+  title = {Maze Generator and Solver - Wolfram Demonstrations Project},
+  year = 2011,
+  url = {https://demonstrations.wolfram.com/MazeGeneratorAndSolver/},
+  urldate = {2020-02-06}
+}  
+
+@article{MazeSuite:2008,
+  title={Maze Suite 1.0: A complete set of tools to prepare, present, and analyze navigational and spatial cognitive neuroscience experiments},
+  author={Ayaz, Hasan and Allen, Sarah L and Platek, Steven M and Onaral, Banu},
+  journal={Behavior research methods},
+  volume={40},
+  number={1},
+  pages={353--359},
+  year={2008},
+  publisher={Springer}
+}
