+++ paper/paper.md
@@ -26,6 +26,20 @@ Mazes have fascinated people since ancient times, and extensive research has bee
 
 A typical problem in describing a new maze-solving algorithm in a journal article is to actually generate the maze image. Although random maze generators are widely available, often a paper needs to describe a new algorithm that optimizes certain types of paths and loops in a maze. In such cases, we need to have precise control over how the maze looks. Moreover, the paths taken by a robot (or solver) also needs to be marked out (e.g. see [@Gupta:2014]). To generate such images, the researcher needs to either painstakingly draw it or write a program to do so, both of which are time-consuming and tedious. Even if such an image could be generated for journal articles, they can not be shared among researchers so that they may try their own solvers on them, because these images cannot be readily converted into program-readable mazes. To solve the aforesaid problems, we have developed a markup language called Maze Description Language (MDL), which is human-readable yet terse.
 
+# State of the field
+
+Existing maze generators can be categorized into four classes:
+
+1. Those which generate mazes as text (e.g. as 0s and 1s, see [@OppenheimJ:2018]).
+
+1. Those which generate random mazes as images/PDF but have no solution paths marked out [@MazeGenerator:2019].
+
+1. Those which generate maze images/document-format and have solution paths marked out [@Wolfram:2011].
+
+1.  Those which generate maze images/3D-models and can have partial paths marked out [@MazeSuite:2008].
+
+The generators in the first category are only suitable as inputs to some maze-solving algorithms and can not be displayed/published in a paper. The second category generators generate only the mazes and no paths can be marked out, making them unsuitable for researchers trying to describe the maze-solving process undertaken by their own proposed algorithms in a step-by-step fashion. The third category generators can only mark out solution paths, but researchers often need to display how some algorithm can take detours, be stuck in dead-ends, etc. The final class of generators is where our proposed tool falls in is suitable for research purposes. However, the ones that do exist currently (such as [@MazeSuite:2008]) are not open-source. Moreover, the paths they depict in MazeViewer (part of MazeSuite) come from log-files generated by researchers. Therefore, the path-traversal data comes purely from physical experiments (with fine resolution) and not suitable for simply depicting path-traversal by an algorithm. Although such path-data (as expected by MazeViewer) can be generated programmtically, it would require more work on the part of the researcher.
+
 # Summary
 
  With MDL-Suite, mazes can be easily described with minimal code, and visited paths can be marked out. Multiple visited paths can also be marked using separate colors. The mazes are fully customizable and a path-based description allows manual maze design very easy. The MDL suite contains two programs: a MDL compiler and a MDL generator. 
