### 2020-02-08 179f2da
Added contribution guidelines
