+++ CONTRIBUTING.md
@@ -1,0 +1,32 @@
+## Contributing to MDL Suite
+
+Thanks for taking the time to contribute! Please read the guidelines below before contributing to MDL Suite.
+
+#### **Reporting a bug**
+
+* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/Akash-Nag/MDL-Suite/issues).
+
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/Akash-Nag/MDL-Suite/issues/new). Be sure to include a **title and clear description**, and as much relevant information as possible. For issues with the MDL compiler, please include a **maze description** on which the issue occurs. Also clearly contrast the expected behavior with the actual output.
+
+#### **Contributing a patch that fixes a bug**
+
+* Open a new GitHub pull request with the patch.
+
+* Ensure the PR description clearly describes the problem and solution. Include the relevant issue number if applicable.
+
+* Before submitting, please ensure all coding conventions have been followed, and that the code is well documented.
+
+#### **Do you intend to add a new feature?**
+
+* Open a new GitHub pull request to merge your fork.
+
+* Clearly describe what the new feature attempts to add to the tool, and why it will be helpful.
+
+* Before submitting, please ensure all coding conventions have been followed, and that the code is well documented.
+
+#### **Do you want to contribute to the documentation?**
+
+* Your contributions are most welcome! Please feel free to contact the author of this repository directly over e-mail to suggest any contributions to the MDL Suite documentation.
+
+
+Thanks!
