### 2020-02-08 eb1efb3
Added contribution guidelines
