### 2020-02-08 Akash-Nag/MDL-Suite#7 @akashnag
I have now added contribution guidelines. Thanks.
