### 2020-02-10 #1815 @drvinceknight
@Akash-Nag it looks like you are missing two DOIs:

- https://doi.org/10.1109/confluence.2014.6949354 may be missing for title: Survey on techniques used in autonomous maze solving robot
- https://doi.org/10.3758/brm.40.1.353 may be missing for title: Maze Suite 1.0: A complete set of tools to prepare, present, and analyze navigational and spatial cognitive neuroscience experiments
