### 2020-02-11 78140fc
Added DOIs to paper
