### 2020-02-11 #1815 @Akash-Nag
I have now added the 2 missing DOIs to the references.

Thanks,
