### 2020-02-11 #1815 @drvinceknight
Sorry @akashnag but you haven't added the DOIs correctly, the DOI for the first one is actually just `10.1109/confluence.2014.6949354 ` here's a full example of a bib file if that's helpful: https://github.com/drvinceknight/Nashpy/blob/master/paper.bib
