### 2020-02-12 d8020f3
Corrected DOIs in references
