+++ paper/paper.bib
@@ -4,7 +4,7 @@
   booktitle={2014 5th International Conference-Confluence The Next Generation Information Technology Summit (Confluence)},
   pages={323--328},
   year={2014},
-  doi={https://doi.org/10.1109/CONFLUENCE.2014.6949354},
+  doi={10.1109/CONFLUENCE.2014.6949354},
   organization={IEEE}
 }
 
@@ -44,6 +44,6 @@
   number={1},
   pages={353--359},
   year={2008},
-  doi={https://doi.org/10.3758/BRM.40.1.353},
+  doi={10.3758/BRM.40.1.353},
   publisher={Springer}
 }
