### 2020-02-13 #1815 @akashnag
I have archived the repository on Zenodo, with version 1.0
Release URL (GitHub): https://github.com/akashnag/MDL-Suite/releases/tag/1.0
DOI: 10.5281/zenodo.3666101
DOI URL: https://doi.org/10.5281/zenodo.3666101
