### 2020-02-17 #1815 @drvinceknight
> @drvinceknight - before you turn this over to the EiCs, please also proof-read the paper and references and make sure you are satisfied with them

Thanks @danielskatz I have actually been doing that but realise that for two recent papers when I handed over to you that you made some subsequent edits. If this is the case here as well I think I'd appreciate a chat where I can understand how I can improve what I'm doing.
