### 2020-02-17 #1815 @drvinceknight
@openjournals/joss-eics this paper is now ready to be accepted. (I believe.)
